elt_doc_rag/pyproject.toml
[project]
name = "elt-doc-rag"
version = "0.1.0"
description = "Configuration-driven document collection management for RAG systems"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "elt-llm-rag>=0.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
elt-doc-rag-ingest = "elt_doc_rag.ingest:main"
elt-doc-rag-query = "elt_doc_rag.query:main"
elt-doc-rag-status = "elt_doc_rag.status:main"

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=4.0.0",
]
